python/requirements.txt
# # backend/requirements.txt
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-multipart==0.0.6
# pydantic==2.5.0

# # Data processing
# numpy==1.24.3
# pandas==2.1.3
# scikit-learn==1.3.2

# # Deep learning
# tensorflow==2.15.0
# torch==2.1.0
# torchvision==0.16.0

# # Explainability
# shap==0.43.0
# lime==0.2.0.1

# # Visualization
# matplotlib==3.8.2
# seaborn==0.13.0
# plotly==5.18.0

# # Image processing
# opencv-python==4.8.1.78
# Pillow==10.1.0

# # Medical imaging
# pydicom==2.4.3

# # Utilities
# python-dotenv==1.0.0
# requests==2.31.0

torch
torchvision
fastapi
uvicorn
numpy
opencv-python
pillow
matplotlib
scikit-learn
python-multipart
